+++ src/actions/LogActions.tsx
@@ -29,6 +29,7 @@ import type {
 import { lstrings } from '../locales/strings'
 import type { ThunkAction } from '../types/reduxTypes'
 import { getCurrencyCode } from '../util/CurrencyInfoHelpers'
+import { base58 } from '../util/encoding'
 import { clearLogs, logWithType, readLogs } from '../util/logger'
 import { getOsVersion } from '../util/utils'
 
@@ -53,13 +54,17 @@ export interface LogOutput {
 }
 
 interface Accounts {
+  /** This is really the "loginId", not the "userId" username hash. */
+  userId: string
+
   username: string
-  userId: string // Not sure what this is used for, but adding the type anyway
 }
 
 interface LoggedInUser {
-  userName: string
+  /** This is really the "loginId", not the "userId" username hash. */
   userId: string
+
+  userName: string
   wallets: WalletData[]
   actions: ActionData[]
 }
@@ -116,7 +121,7 @@ export function showClearLogsModal(): ThunkAction<Promise<void>> {
           }
         }}
       />
-    )).catch(error => {
+    )).catch((error: unknown) => {
       showError(error)
     })
   }
@@ -153,12 +158,12 @@ export function getLogOutput(): ThunkAction<Promise<MultiLogOutput>> {
     const { actionQueue, core } = state
     const { account, context } = core
 
-    if (context) {
+    if (context != null) {
       // Get local accounts
       for (const user of context.localUsers) {
         logOutput.accounts.push({
-          username: user.username ?? '<undefined>',
-          userId: ''
+          userId: base64.stringify(base58.parse(user.loginId)),
+          username: user.username ?? '<undefined>'
         })
       }
     }
@@ -170,7 +175,7 @@ export function getLogOutput(): ThunkAction<Promise<MultiLogOutput>> {
       const { syncKey } = await account.getRawPrivateKey(account.id)
 
       logOutput.loggedInUser = {
-        userId: rootLoginId,
+        userId: base64.stringify(base58.parse(rootLoginId)),
         userName: username ?? '<undefined>',
         wallets: [],
         actions: []
@@ -201,7 +206,7 @@ export function getLogOutput(): ThunkAction<Promise<MultiLogOutput>> {
         const { imported, syncKey } = await account.getRawPrivateKey(wallet.id)
 
         // Wallet info
-        if (wallet && logOutput.loggedInUser) {
+        if (wallet != null && logOutput.loggedInUser != null) {
           const currencyCode = wallet.currencyInfo.currencyCode ?? ''
           logOutput.loggedInUser.wallets.push({
             created: wallet.created?.toISOString(),
@@ -209,7 +214,7 @@ export function getLogOutput(): ThunkAction<Promise<MultiLogOutput>> {
             customTokens: wallet.currencyConfig.customTokens,
             imported,
             repoId: getRepoId(syncKey),
-            pluginDump: await wallet.dumpData().catch(error => ({
+            pluginDump: await wallet.dumpData().catch((error: unknown) => ({
               walletId: wallet.id,
               walletType: wallet.type,
               data: { dumpError: { message: String(error) } }
@@ -257,14 +262,17 @@ function getRepoId(key: string): string {
   return 'Invalid syncKey type'
 }
 
-export const sendLogs = async (logs: LogOutput, underDuress: boolean) => {
+export const sendLogs = async (
+  logs: LogOutput,
+  underDuress: boolean
+): Promise<void> => {
   console.log('====== SENDING LOGS REQUEST ======')
 
   if (underDuress) {
     logs.userMessage += ' (DURESS MODE)'
   }
 
-  return await fetch(logsUri, {
+  await fetch(logsUri, {
     method: 'PUT',
     headers: {
       Accept: 'application/json',
@@ -279,7 +287,7 @@ export const sendLogs = async (logs: LogOutput, underDuress: boolean) => {
       console.log(`====== SENDING LOGS SUCCESS ======`, response)
       return response
     })
-    .catch(error => {
+    .catch((error: unknown) => {
       console.log(`====== SENDING LOGS FAILURE ======`, error)
       throw error
     })