+++ src/constants/WalletAndCurrencyConstants.ts
@@ -167,6 +167,7 @@ interface SpecialCurrencyInfo {
   noChangeMiningFee?: boolean
   noMaxSpend?: boolean
   keysOnlyMode?: boolean
+  highPrecisionSyncRatioDisplay?: boolean
   /**
    * This disables the transaction list for the wallet.
    * (Default: false)
@@ -200,7 +201,7 @@ interface SpecialCurrencyInfo {
 export const getSpecialCurrencyInfo = (
   pluginId: string
 ): SpecialCurrencyInfo => {
-  if (SPECIAL_CURRENCY_INFO[pluginId]) {
+  if (SPECIAL_CURRENCY_INFO[pluginId] !== undefined) {
     return SPECIAL_CURRENCY_INFO[pluginId]
   } else {
     return {
@@ -832,6 +833,7 @@ export const SPECIAL_CURRENCY_INFO: Record<string, SpecialCurrencyInfo> = {
     noChangeMiningFee: true,
     isImportKeySupported: true,
     keysOnlyMode: Platform.OS === 'android' && Platform.constants.Version < 28,
+    highPrecisionSyncRatioDisplay: true,
     importKeyOptions: [
       {
         optionName: 'birthdayHeight',
@@ -973,7 +975,7 @@ export const USD_FIAT = 'iso:USD'
 /**
  * Get the fiat symbol from an iso:[fiat] OR fiat currency code
  */
-export const getFiatSymbol = (isoOrFiatCurrencyCode: string) => {
+export const getFiatSymbol = (isoOrFiatCurrencyCode: string): string => {
   if (typeof isoOrFiatCurrencyCode !== 'string') return ''
   const codeWithoutIso = removeIsoPrefix(isoOrFiatCurrencyCode)
   const out = FIAT_CODES_SYMBOLS[codeWithoutIso.toUpperCase()]