+++ CHANGELOG.md
@@ -2,8 +2,17 @@
 
 ## Unreleased (develop)
 
-## 4.38.0 (staging)
+## 4.39.0 (staging)
 
+- added: BitsOfGold and LibertyX ramp plugins.
+- added: Added remote ramp plugin filtering from info-server rollup.
+- fixed: Fixed broken `logEvent` tracking calls by adding the needed `dispatch`.
+- fixed: Localized display names for payment types in new buy/sell feature.
+- removed: Remove change quote tracking.
+
+## 4.38.0 (2025-11-04)
+
+- added: Include loginId's for all accounts when uploading logs.
 - changed: Reduced number of views in `WalletListCurrencyRow`
 - changed: Custom assets to show a generic grey icon, overlayed with the first three letters of the asset
 - changed: "X" button re-added to `EdgeModal` for accessibility